requirements.txt
google-generativeai==0.3.0
neo4j==5.14.0
torch==2.2.0
transformers==4.37.0
networkx==3.2.1
matplotlib==3.8.2
scipy==1.12.0
ttkbootstrap==1.10.1
customtkinter==5.2.1
